Dither Fixes

+++ Scripts/PSX/PSXEffectFeature.cs
@@ -43,10 +43,10 @@ namespace ColbyO.VNTG.PSX
 
             mat.SetFloat("_ColorPrecision", settings.ColorPrecision.value);
 
-            mat.SetInt("_EnableDither", (settings.EnableDither.value) ? 1 : 0);
-            mat.SetInt("_DitherMode", (int)settings.DitherMode.value);
+            mat.SetFloat("_EnableDither", (settings.EnableDither.value) ? 1 : 0);
+            mat.SetFloat("_DitherMode", (int)settings.DitherMode.value);
             mat.SetInt("_DitherPattern", settings.DitherPattern.value);
-            mat.SetInt("_DitherPixelPerfect", settings.DitherPixelPerfect.value ? 1 : 0);
+            mat.SetFloat("_DitherPixelPerfect", settings.DitherPixelPerfect.value ? 1 : 0);
             mat.SetFloat("_DitherScale", Mathf.Lerp(1f, 10f, settings.DitherScale.value));
             mat.SetFloat("_DitherThreshold", settings.DitherThreshold.value);
